producer/requirements.txt
confluent-kafka >= 2.2.0
python-dateutil >= 2.8.2 